feat!: trying to organize the whole dotfile setup
- Move unused scripts to `legacy/`
- Move used scrips to either `frontend/` or `backend/`.
- Fix `volumectl.sh` not properly showing icons in notifications.
- Add `special:resmon` to ignored workspaces in `get_client_quantity.py`.
- Use `bluetooth-manager` instead of Overskride.
- Use `jq` in acquiring the window geometry in `screenshot.sh`.
- In Swaylock, do not lock the screen if the user cancelled the prompt.
- Call `resource-monitor.sh` instead of running `kitty btop` when CPU is
  click on Waybar.
- Ask for confirmation in sleep, logout, restart, and poweroff in
  `powermenu.sh`.

+++ scripts/frontend/powermenu.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
-ROFI_THEME="$HOME/.config/rofi/config/powermenu.rasi"
+MENU_ROFI_THEME="$HOME/.config/rofi/config/powermenu.rasi"
+CONF_ROFI_THEME="$HOME/.config/rofi/config/confirm.rasi"
+
+yes=""
+no=""
 
 lock=""
 sleep=""
@@ -8,22 +12,43 @@ logout=""
 restart="󰜉"
 poweroff="󰐥"
 
-user_selection=$(printf "%s\n%s\n%s\n%s\n%s\n" "$lock" "$sleep" "$logout" "$restart" "$poweroff" | rofi -dmenu -theme "$ROFI_THEME")
+user_selection=$(printf "%s\n%s\n%s\n%s\n%s\n" "$lock" "$sleep" "$logout" "$restart" "$poweroff" | rofi -dmenu -theme "$MENU_ROFI_THEME")
 
 case "$user_selection" in
 "$lock")
+    # This script has its own confirmation prompt.
     ~/.config/scripts/frontend/lockscreen.sh
     ;;
 "$sleep")
-    ~/.config/scripts/frontend/lockscreen.sh --no-private && systemctl suspend
+    sleep_confirm=$(printf "%s\n%s" "$no" "$yes" | rofi -dmenu -i -p "Do you really want to suspend the system?" -theme "$CONF_ROFI_THEME")
+    if [ -z "$sleep_confirm" ]; then
+        exit 1
+    elif [ "$sleep_confirm" = "$yes" ]; then
+        ~/.config/scripts/frontend/lockscreen.sh --no-private && systemctl suspend
+    fi
     ;;
 "$logout")
-    hyprctl dispatch exit
+    logout_confirm=$(printf "%s\n%s" "$no" "$yes" | rofi -dmenu -i -p "Do you really want to log out?" -theme "$CONF_ROFI_THEME")
+    if [ -z "$logout_confirm" ]; then
+        exit 1
+    elif [ "$logout_confirm" = "$yes" ]; then
+        hyprctl dispatch exit
+    fi
     ;;
 "$restart")
-    systemctl reboot
+    restart_confirm=$(printf "%s\n%s" "$no" "$yes" | rofi -dmenu -i -p "Do you really want to restart the system?" -theme "$CONF_ROFI_THEME")
+    if [ -z "$restart_confirm" ]; then
+        exit 1
+    elif [ "$restart_confirm" = "$yes" ]; then
+        systemctl reboot
+    fi
     ;;
 "$poweroff")
-    systemctl poweroff
+    poweroff_confirm=$(printf "%s\n%s" "$no" "$yes" | rofi -dmenu -i -p "Do you really want to shut down the system?" -theme "$CONF_ROFI_THEME")
+    if [ -z "$poweroff_confirm" ]; then
+        exit 1
+    elif [ "$poweroff_confirm" = "$yes" ]; then
+        systemctl poweroff
+    fi
     ;;
 esac
